+++ shopApp/app/app/src/main/java/com/example/shopping/models/Products.kt
@@ -3,7 +3,7 @@ package com.example.shopping.models
 data class ProductModel(
     val id: String,
     val name: String,
-    val price: Double,
+    val price: Float,
     val category: CategoryModel,
     val details: String
 )
